Enhances chat message rendering
- Added code block header with language + title
- Added copy code block to clipboard button
- Changed user message to a right-aligned chat bubble with no avatar
- Made various chat message layout modifications
- Enabled bottom bar for user messages (copy to clipboard only)
