+++ frontend/package.json
@@ -38,6 +38,7 @@
     "react-router-dom": "^6.25.1",
     "react-syntax-highlighter": "^15.5.0",
     "reactflow": "^11.11.4",
+    "remark-code-title": "^0.2.5",
     "remark-gfm": "^4.0.0",
     "remark-github-blockquote-alert": "^1.2.1",
     "tailwindcss": "^3.4.6",
