+++ frontend/src/components/chat/Mentions.tsx
@@ -82,7 +82,7 @@ export const ExplanationMention: FC<{
   }
 
   return (
-    <div className="flex flex-wrap items-center gap-2 py-2">
+    <div className="flex flex-wrap items-center gap-2 pt-2">
       <span
         className={"text-xs font-semibold text-gray-900 dark:text-gray-500"}
       >
